Allow decimal value for number of seconds in -l, because why not?

+++ src/main.cpp
@@ -113,7 +113,7 @@ int main(int argc, char** argv) {
     // appropriate output file is set.
     if (!limit_arg.empty()) {
       // Internally timeout is in milliseconds.
-      cl_args.timeout = 1000 * std::stoul(limit_arg);
+      cl_args.timeout = 1000 * std::stof(limit_arg);
     }
     cl_args.nb_threads = std::stoul(nb_threads_arg);
     cl_args.exploration_level = std::stoul(exploration_level_arg);
